+++ src/app/latte-chat/mypage/edit/page.tsx
@@ -1,6 +1,4 @@
-'use client'
-
-import MypageEditContainer from './MypageEditContainer'
+import MypageEditContainer from '@/features/user-mypage-edit/containers/MypageEditContainer'
 
 export default function MypageEditPage() {
   return <MypageEditContainer />
